finished resource write web service
git-svn-id: https://svn.code.sf.net/p/phpscheduleit/code/development@1846 12d54735-0ee2-4768-9b18-c71e5e561e4c

+++ WebServices/Controllers/ResourceSaveController.php
@@ -104,6 +104,11 @@ class ResourceSaveController implements IResourceSaveController
 	 */
 	public function Delete($resourceId, $session)
 	{
+		$errors = $this->validator->ValidateDeleteRequest($resourceId);
+		if (!empty($errors))
+		{
+			return new ResourceControllerResult(null, $errors);
+		}
 		$resource = $this->repository->LoadById($resourceId);
 		$this->repository->Delete($resource);
 
@@ -141,6 +146,14 @@ class ResourceSaveController implements IResourceSaveController
 		}
 		$resource->ChangeAttributes($attributes);
 
+		if ($request->isOnline)
+		{
+			$resource->BringOnline();
+		}
+		else
+		{
+			$resource->TakeOffline();
+		}
 		return $resource;
 	}
 }
